modify registerForm
