+++ conscious-program.iml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module version="4">
   <component name="SonarLintModuleSettings">
-    <option name="uniqueId" value="680eda67-1c0d-47b3-a658-94fc8ce6d34e" />
+    <option name="uniqueId" value="92d4282f-814f-4769-9a40-5458d518bccf" />
   </component>
 </module>
