--- swing/JavaLogIn.form
@@ -1,150 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="JavaLogIn">
-  <grid id="27dc6" binding="JavaLogInPanel" layout-manager="GridLayoutManager" row-count="9" column-count="3" same-size-horizontally="false" same-size-vertically="false" hgap="0" vgap="0">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="4" y="5" width="570" height="675"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <component id="f370f" class="javax.swing.JButton" binding="submitButton" default-binding="true">
-        <constraints>
-          <grid row="8" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <background color="-852843"/>
-          <text value="Register"/>
-        </properties>
-      </component>
-      <component id="fcd4e" class="javax.swing.JLabel" binding="javaTitle">
-        <constraints>
-          <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false">
-            <preferred-size width="462" height="214"/>
-          </grid>
-        </constraints>
-        <properties>
-          <font name="Source Code Pro" size="22" style="1"/>
-          <horizontalAlignment value="2"/>
-          <horizontalTextPosition value="11"/>
-          <icon value="elac-logo.png"/>
-          <text value="Java Log-In Form"/>
-        </properties>
-      </component>
-      <component id="8fc65" class="javax.swing.JTextField" binding="textFirstName">
-        <constraints>
-          <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties/>
-      </component>
-      <component id="e5c49" class="javax.swing.JLabel" binding="javaAddress">
-        <constraints>
-          <grid row="4" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="1" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="Address:"/>
-        </properties>
-      </component>
-      <component id="63c99" class="javax.swing.JLabel" binding="javaPhone">
-        <constraints>
-          <grid row="3" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="  Phone # :"/>
-        </properties>
-      </component>
-      <component id="6e297" class="javax.swing.JFormattedTextField" binding="textPhone">
-        <constraints>
-          <grid row="3" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties/>
-      </component>
-      <component id="ca7df" class="javax.swing.JTextField" binding="textAddress">
-        <constraints>
-          <grid row="4" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties/>
-      </component>
-      <component id="6a814" class="javax.swing.JLabel" binding="javaFirstName">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="  First Name:"/>
-        </properties>
-      </component>
-      <component id="ca42c" class="javax.swing.JLabel" binding="javaLastName">
-        <constraints>
-          <grid row="2" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <opaque value="true"/>
-          <requestFocusEnabled value="true"/>
-          <text value=" Last Name:"/>
-        </properties>
-      </component>
-      <component id="71b50" class="javax.swing.JTextField" binding="textLastName">
-        <constraints>
-          <grid row="2" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties/>
-      </component>
-      <component id="6264d" class="javax.swing.JLabel" binding="javaCity">
-        <constraints>
-          <grid row="5" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="  City:"/>
-        </properties>
-      </component>
-      <component id="dbecf" class="javax.swing.JTextField" binding="textCity">
-        <constraints>
-          <grid row="5" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="100" height="-1"/>
-          </grid>
-        </constraints>
-        <properties/>
-      </component>
-      <component id="a7221" class="javax.swing.JLabel" binding="javaZipcode">
-        <constraints>
-          <grid row="6" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value="  Zipcode:"/>
-        </properties>
-      </component>
-      <component id="910c6" class="javax.swing.JTextField" binding="textZipcode">
-        <constraints>
-          <grid row="6" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties/>
-      </component>
-      <component id="c44b0" class="javax.swing.JLabel" binding="javaUserName">
-        <constraints>
-          <grid row="7" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <text value=" User Name:"/>
-        </properties>
-      </component>
-      <component id="b9834" class="javax.swing.JTextField" binding="textUsername">
-        <constraints>
-          <grid row="7" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="8" fill="1" indent="0" use-parent-layout="false">
-            <preferred-size width="150" height="-1"/>
-          </grid>
-        </constraints>
-        <properties/>
-      </component>
-    </children>
-  </grid>
-</form>
