+++ swing/src/main/java/org/elac/SwingApp.java
@@ -2,7 +2,7 @@ package org.elac;
 
 import java.awt.EventQueue;
 import javax.swing.JFrame;
-import org.elac.form.JavaLogIn;
+import org.elac.form.RegisterForm;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
@@ -13,7 +13,7 @@ import org.springframework.context.ConfigurableApplicationContext;
 public class SwingApp extends JFrame {
 
     public SwingApp() {
-        new JavaLogIn();
+        new RegisterForm();
     }
